fix: refine a little bit
