+++ packages/draft-editor/src/block-renderers/image-link.tsx
@@ -1,13 +1,47 @@
 import React, { useState } from 'react'
 import styled from 'styled-components'
-import { AtomicBlockProps } from '../block-renderer-fn.type'
+import { Drawer, DrawerController } from '@keystone-ui/modals'
+import { TextInput } from '@keystone-ui/fields'
 import {
-  ImageSelectorOnChangeFn,
-  ImageSelector,
-  ImageEntityWithMeta,
-} from '../buttons/selector/image-selector'
-import { EditableBlock as _EditableBlock } from './styled'
+  EditorState,
+  RawDraftContentState,
+  convertToRaw,
+  convertFromRaw,
+} from 'draft-js'
 import { blockRenderers } from '@kids-reporter/draft-renderer'
+import { AtomicBlockProps } from '../block-renderer-fn.type'
+import { ImageEntityWithMeta } from '../buttons/selector/image-selector'
+import { EditableBlock as _EditableBlock } from './styled'
+import { RichTextEditor } from '../rich-text-editor'
+import { editableLinkDecorator } from '../entity-decorators/link'
+import buttonNames from '../buttons/bt-names'
+
+const disabledButtons = [
+  buttonNames.bold,
+  buttonNames.italic,
+  buttonNames.underline,
+  buttonNames.fontColor,
+  buttonNames.backgroundColor,
+  buttonNames.ol,
+  buttonNames.ul,
+  buttonNames.divider,
+  buttonNames.h2,
+  buttonNames.h3,
+  buttonNames.h4,
+  buttonNames.h5,
+  buttonNames.code,
+  buttonNames.codeBlock,
+  buttonNames.blockquote,
+  buttonNames.infoBox,
+  buttonNames.slideshow,
+  buttonNames.newsReading,
+  buttonNames.tocAnchor,
+  buttonNames.anchor,
+  buttonNames.embed,
+  buttonNames.annotation,
+  buttonNames.image,
+  buttonNames.imageLink,
+]
 
 const { ImageLinkInArticleBody } = blockRenderers
 
@@ -25,6 +59,73 @@ type EntityData = ImageEntityWithMeta & {
   alignment?: string
 }
 
+export type ImageLinkValue = {
+  type: 'image-link'
+  rawContentState: RawDraftContentState
+}
+
+export const ImageLinkEditor = (props: {
+  isOpen: boolean
+  inputValue: ImageLinkValue
+  onConfirm: (arg0: {
+    type: 'image-link'
+    rawContentState: RawDraftContentState
+  }) => void
+  onCancel: () => void
+}) => {
+  const { isOpen, inputValue, onConfirm, onCancel } = props
+  const [url, setURL] = useState('')
+  const contentState = convertFromRaw(inputValue.rawContentState)
+  const [inputValueState, setInputValueState] = useState({
+    type: inputValue.type,
+    editorState: EditorState.createWithContent(contentState),
+  })
+
+  console.log(setInputValueState)
+
+  return (
+    <DrawerController isOpen={isOpen}>
+      <Drawer
+        title={`Image Link`}
+        actions={{
+          cancel: {
+            label: 'Cancel',
+            action: () => onCancel(),
+          },
+          confirm: {
+            label: 'Confirm',
+            action: () =>
+              onConfirm({
+                type: inputValueState.type,
+                rawContentState: convertToRaw(
+                  inputValueState.editorState.getCurrentContent()
+                ),
+              }),
+          },
+        }}
+      >
+        <TextInput
+          placeholder="圖片連結"
+          type="text"
+          value={url}
+          onChange={(e) => setURL(e.target.value)}
+        />
+        <RichTextEditor
+          decorators={[editableLinkDecorator]}
+          disabledButtons={disabledButtons}
+          editorState={inputValueState.editorState}
+          onChange={(editorState: EditorState) => {
+            setInputValueState({
+              type: inputValueState.type,
+              editorState,
+            })
+          }}
+        />
+      </Drawer>
+    </DrawerController>
+  )
+}
+
 export const EditableImageLink = (props: AtomicBlockProps<EntityData>) => {
   const [isSelectorOpen, setIsSelectorOpen] = useState(false)
   const { block, blockProps, contentState } = props
@@ -34,7 +135,9 @@ export const EditableImageLink = (props: AtomicBlockProps<EntityData>) => {
   const data = entity.getData() || {}
   const {alignment: _alignment, ...imageWithMeta} = data // eslint-disable-line
 
-  const onChange: ImageSelectorOnChangeFn = (selectedImages, alignment) => {
+  const onChange = () => {
+    onEditFinish()
+    /*
     setIsSelectorOpen(false)
 
     if (selectedImages?.length === 0) {
@@ -48,17 +151,22 @@ export const EditableImageLink = (props: AtomicBlockProps<EntityData>) => {
       entityKey,
       entityData: Object.assign({ alignment: alignment }, selectedImage),
     })
+    */
   }
 
   return (
     <>
       {isSelectorOpen && (
-        <ImageSelector
-          onChange={onChange}
-          enableCaption={true}
-          enableAlignment={true}
-          alignment={data.alignment}
-          selected={[imageWithMeta]}
+        <ImageLinkEditor
+          isOpen={isSelectorOpen}
+          inputValue={{
+            type: 'image-link',
+            rawContentState: { blocks: [], entityMap: {} },
+          }}
+          onConfirm={onChange}
+          onCancel={() => {
+            setIsSelectorOpen(false)
+          }}
         />
       )}
       <EditableBlock
